second commit

+++ src/main/java/com/example/demo/repository/TransactionRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Transactions;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepository extends JpaRepository<Transactions,Integer> {
+public interface TransactionRepository extends JpaRepository<Transactions,Long> {
 }
